Move main csbot command to csbot.cli module
