+++ src/csbot/util.py
@@ -1,6 +1,6 @@
 import shlex
 from itertools import tee
-from collections import deque, OrderedDict
+from collections import deque
 import asyncio
 import logging
 import typing
@@ -361,3 +361,53 @@ def type_validator(_obj, attrib: attr.Attribute, value):
         return True
     raise TypeError(f"'{attrib.name}' must be {attrib.type} (got {value} that is a {type(value)}",
                     attrib, value)
+
+
+class PrettyStreamHandler(logging.StreamHandler):
+    """Wrap log messages with severity-dependent ANSI terminal colours.
+
+    Use in place of :class:`logging.StreamHandler` to have log messages coloured
+    according to severity.
+
+    >>> handler = PrettyStreamHandler()
+    >>> handler.setFormatter(logging.Formatter('[%(levelname)-8s] %(message)s'))
+    >>> logging.getLogger('').addHandler(handler)
+
+    *stream* corresponds to the same argument to :class:`logging.StreamHandler`,
+    defaulting to stderr.
+
+    *colour* overrides TTY detection to force colour on or off.
+
+    This source for this class is released into the public domain.
+
+    .. codeauthor:: Alan Briolat <alan.briolat@gmail.com>
+    """
+    #: Mapping from logging levels to ANSI colours.
+    COLOURS = {
+        logging.DEBUG: '\033[36m',      # Cyan foreground
+        logging.WARNING: '\033[33m',    # Yellow foreground
+        logging.ERROR: '\033[31m',      # Red foreground
+        logging.CRITICAL: '\033[31;7m'  # Red foreground, inverted
+    }
+    #: ANSI code for resetting the terminal to default colour.
+    COLOUR_END = '\033[0m'
+
+    def __init__(self, stream=None, colour=None):
+        super(PrettyStreamHandler, self).__init__(stream)
+        if colour is None:
+            self.colour = self.stream.isatty()
+        else:
+            self.colour = colour
+
+    def format(self, record):
+        """Get a coloured, formatted message for a log record.
+
+        Calls :func:`logging.StreamHandler.format` and applies a colour to the
+        message if appropriate.
+        """
+        msg = super(PrettyStreamHandler, self).format(record)
+        if self.colour:
+            colour = self.COLOURS.get(record.levelno, '')
+            return colour + msg + self.COLOUR_END
+        else:
+            return msg
